feat: heartbeat and only send position if needed
closes: #10

+++ server/include/Server.h
@@ -36,7 +36,7 @@ class Server : public QObject {
   Q_OBJECT
 
  public:
-  explicit Server(quint8 port);
+  explicit Server(int port);
 
  private slots:
   void clientConnect();
